+++ packages/bluesky/test/src/feeds/feeds_service_test.dart
@@ -9,6 +9,7 @@ import 'package:atproto_test/atproto_test.dart' as atp_test;
 import 'package:bluesky/src/entities/actor_feeds.dart';
 import 'package:bluesky/src/entities/feed.dart';
 import 'package:bluesky/src/entities/feed_generator.dart';
+import 'package:bluesky/src/entities/feed_generator_info.dart';
 import 'package:bluesky/src/entities/feed_generators.dart';
 import 'package:bluesky/src/entities/likes.dart';
 import 'package:bluesky/src/entities/post_thread.dart';
@@ -1641,6 +1642,68 @@ void main() {
         () async => await feeds.findGenerators(
           uris: [],
         ),
+      );
+    });
+  });
+
+  group('.findGeneratorInfo', () {
+    test('normal case', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/feeds/data/find_generator_info.json',
+        ),
+      );
+
+      final response = await feeds.findGeneratorInfo();
+
+      expect(response, isA<XRPCResponse>());
+      expect(response.data, isA<FeedGeneratorInfo>());
+    });
+
+    test('when unauthorized', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await feeds.findGeneratorInfo(),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await feeds.findGeneratorInfo(),
       );
     });
   });