+++ packages/bluesky_cli/lib/src/command/commands.dart
@@ -17,6 +17,7 @@ import 'feed/custom_feed.dart';
 import 'feed/feed.dart';
 import 'feed/feed_generator.dart';
 import 'feed/feed_generators.dart';
+import 'feed/generator_info.dart';
 import 'feed/like.dart';
 import 'feed/likes.dart';
 import 'feed/post.dart';
@@ -72,6 +73,7 @@ final feedCommands = <BskyCommand>[
   FeedGeneratorCommand(),
   FeedGeneratorsCommand(),
   CustomFeedCommand(),
+  GeneratorInfoCommand(),
 ];
 
 final notificationCommands = <BskyCommand>[